+++ packages/remix-testing/CHANGELOG.md
@@ -1,5 +1,21 @@
 # `@remix-run/testing`
 
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/react@2.8.1`
+  - `@remix-run/node@2.8.1`
+
+## 2.8.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/react@2.8.0`
+  - `@remix-run/node@2.8.0`
+
 ## 2.7.2
 
 ### Patch Changes